refactor: Chat 관련 로직을 Room 하위 패키지로 이동
